Use quotes where appropriate and prefer bash [[
[[ is a shell keyword, in which parameter expansion is not subject to
shell word-splitting, unlike in the built-in [ (or even /bin/[). It is
generally faster and more powerful, supports compound booleans, and
should be used everywhere [ would be, unless writing for #!/bin/sh.

Outside of [[ context, parameter expansion should be done within
double-quotes unless word-splitting is expected and desired, such as
$JAVA_OPTS in line 47.

Curly-braces around parameter expansion are not required anywhere in this
script, but was left in as a helpful visual delimiter in strings where
parameters and text are not separated by whitespace.